requirements.txt
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
huggingface_hub>=0.17.0
rasterio>=1.3.0
Pillow>=10.0.0
PyYAML>=6.0
numpy>=1.24.0
scikit-image>=0.21.0
matplotlib>=3.7.0
opencv-python-headless>=4.8.0
requests>=2.31.0
psutil>=5.9.0